pytransifex/config-plugins/opengis-mkdocs/requirements.txt
python-frontmatter==1.0.0
